Reset vpaths after CreateProject has been called
Added premake project lua files to solutions
Fixed a couple instances where vpath was not setup correctly
Refactored MetaGen to now run within our Premake Build System
Removed TypeParser, TypeGenerator

+++ Source/Network/Network/Define.h
@@ -1,8 +1,6 @@
 #pragma once
 #include <Base/Types.h>
 
-#include <Meta/Generated/Shared/NetworkEnum.h>
-
 #include <memory>
 
 class Bytebuffer;
@@ -18,7 +16,7 @@ namespace Network
         
         { t.Serialize(buffer) } -> std::same_as<bool>;
         { t.Deserialize(buffer) } -> std::same_as<bool>;
-        { t.GetSerializedSize() } -> std::same_as<u16>;
+        { t.GetSerializedSize() } -> std::same_as<u32>;
     };
 
     struct ConnectionInfo
